add configure option to use size_t for length params
The use of unsigned long long is completely bogus.  Unfortunately fixing
this is an incompatible API change, therefore it has to be explicitly
requested via ./configure --enable-size_t

A typedef has been added, sodium_size_t, to allow code to work with both
variants of the API.  Users of libsodium should be encouraged to make
use of it, to hopefully allow the default to change in the future.